Update to use Identity Credential Jetpack.
Instead of using a local copy, use the Jetpack now that it's been published. The API is a bit
different so update the code accordingly. Also update both holder and reader to work when
a MAC is returned instead of an ECDSA signature (this happens when the Jetpack calls into
the underlying Identity Credential HAL).