Preserve order when replacing NULL in IssuerSignedItemBytes CBOR.
The helper Util.issuerSignedItemBytesSetValue() takes the bytes of
`IssuerSignedItemBytes` and replaces the value (usually NULL) for the
"elementValue" key with a given value. It then returns the bytes of
`IssuerSignedItemBytes`.

Because `IssuerSignedItemBytes` originates from the issuer (as part of
`staticAuthData`, see Utility.encodeStaticAuthData() for the CDDL) and
`IssuerSignedItem` is a map, it's not well-defined what the encoding
is. The issuer could choose whatever order they want.

This change fixes Util.issuerSignedItemBytesSetValue() so the order is
preserved. This is done by upgrading cbor-java to 0.9 and using the
non-canonical encoder. Also add unit test for this.

Test: All unit tests pass.
Test: Manually tested holder and verifier.