Refactor occurrences of new Thread(new Runnable() { ...
Instead of constructing an instance of an anonymous subclass of Runnable, constructing an anonymous subclass of Thread and overriding its run() method has the same effect, is more concise, and saves one Object. I'm not aware of any advantage of using the other variant. Therefore this commit refactors such occurrences.