CredentialData: Fix misnamed variable confusing time units.
Before this CL, CredentialData.createCredentialData() called ensureAcpTimoutKeyForProfile(..., timeoutSeconds) but that method expects a timeout in milliseconds.

Upon further inspection, timeoutSeconds came from profile.getUserAuthenticationTimeout(); that method did not document its time unit, but the corresponding Builder method is clear about it being milliseconds. Therefore it appears that the logic is correct but the name 'timeoutSeconds' was wrong.

This commit renames the misnamed variable and adds documentation to AccessControlProfile.getUserAuthenticationTimeout() noting that the returned time value is in milliseconds. I also renamed AccessControlProfile's field mUserAuthenticationTimeout -> mUserAuthenticationTimeoutMillis.

This commit is a pure refactoring with no behavior change.