Add IdentityCredentialTest. (#65)
Add IdentityCredentialTest.

For now, the test covers storing of static authentication data of varying size for varying numbers of authKeys. The reason this is worth testing is that storing this breaks a subsequent getCredentialByName() specifically for large data (100 KBytes+ across keys), apparently on API <= 28. For now this test is disabled via Assume.assumeTrue() on API versions where it is known to break, but this might be tweaked in a future commit if/when we better understand the exact root cause and conditions that trigger it.

I've confirmed that without the Assume.assumeTrue(), the test would break on the following emulated device / API combinations:

 - Pixel 2 API 24
 - Pixel 2 API 26
 - Pixel 3 API 28

It passes on:

 - Pixel 3 API 29
 - Pixel 5 API 31

The test is parameterized to cover all IdentityCredentialStore implementations available on the device. The test runs much slower on the software instance: On the Pixel 5 API 31 emulator on my laptop, it runs for 7 seconds and 15 seconds for the smaller and larger set of test data, respectively.