Switch egress hardening to audit mode
One of the endpoints we need is part of a Microsoft-hosted CDN, and we
get bounced to different servers in this domain depending on the whims
of the load balancer. Thus, we cannot allow-list the exact server,
but harden-runner also doesn't support wildcards. Just use audit mode
for now so we can at least see what's being accessed in the logs.