Partially break dependency on android.icu
We want to allow this code to run on the JVM as well as in Android.
This means we need to avoid dependencies on Android-only classes like
the android.icu.* namespace.

The main use of android.icu is the Calendar class. Replace all internal
uses with a new Timestamp class. Public APIs still use Calendar for
now, as we don't want to break API compatibility just yet.