Fix double-wrapping of IssuerSignedItem.
The root problem is that both encodeStaticAuthData() and
decodeStaticAuthData() have a bug where IssuerSignedItem is wrapped in
a #6.24-tagged bstr twice.

For self-signed credentials provisioned on the device via
Utility.provisionSelfSignedCredential() this turned out to not be a
problem because the two bugs cancel each other out.

However for applications where the StaticAuthData (e.g. MSO and the
digest id mapping) is generated outside the device - the usual case
for production-level mDLs - this ends up being a problem because the
generated StaticAuthData CBOR fails to decode because of the bug in
decodeStaticAuthData().

Fix this by slightly reworking the API so IssuerSignedItem blobs are
passed around, not IssuedSignedItemBytes. Specifically this affects
the following public APIs:

 - Utility.mergeIssuerSigned()
 - Utility.decodeStaticAuthData()
 - Utility.encodeStaticAuthData()
 - DeviceResponseGenerator.addDocument()

This change won't affect the common path where the mdoc application is
just using DeviceResponseGenerator.addDocument() together with
Utility.decodeStaticAuthData().

Additionally, since this is a bug in how StaticAuthData is encoded and
this is stored on disk, you'll need to uninstall the reference app and
install it again to force creation of a new self-signed credential.

Test: New unit tests and all unit tests pass.
Test: Manually tested with reference apps.