Handle the case where no biometrics are enrolled on the device.
When using the Keystore-backed implementation, it's not possible to
create a credential unless biometrics are enrolled. This is fixed by
just avoiding creating keys with timeout 0 since they are not needed
anyway.

Also rework the reference mDL app so it actually uses user
authentication and rework the code so it uses the LSKF if no
biometrics are enrolled on the device. Also add a "Add PIN" button so
the user can always opt to use the LSKF.

Carefully test this on devices with fingerprint (Pixel 3XL) and face
authentication (Pixel 4XL) and the paths where the credential is
created when a biometric is present and when it's not present. Also
test presentation in both modes and the mode where the credential was
created when a biometric was present but subsequently removed.

Test: See above and all unit tests pass.
Bug: 217171229