Rename Software-backed implementation and improve API.
Referring to the Hardware-Backed Android Keystore implementation of
IdentityCredentialStore as "software" is unhelpful and
misleading. Instead refer to this as "Hardware-Backed Keystore" and
update the API to get such a store to use this terminology.

Also provide a way for the application to specify where to store
credential files when using the Keystore-backed implementation. Also
point out that the application should be careful to store this in a
location that is not backed up and update the reference app to do so
Also update the name of the app's document DB to force recreation of
credentials.

Remove IdentityCredentialStore.getDefaultInstance() as it's trivial to
implement on the application side and doesn't easily allow to pass
per-implementation parameters as we now have for the Keystore-backed
implementation.

Remove IdentityCredentialStoreCapabilities class as it's not useful
since applications are expected to be written to support a minimum
feature version and fall back to the KS-store if this requirement is
not met.

Rename IdentityCredentialTest to StoreStaticAuthenticationDataTest and
remove the parametrization since the test suite is specifically
intended to only run against the Keystore implementation (the
HW-backed implementation has its own test suite).

Test: All unit tests pass and manually tested.
Bug: None