BLE L2CAP: defer closing socket to work around possible Bluetooth bug.
When closing a data transport using BLE L2CAP, we send SessionData
with empty payload and status 20 (9 bytes), flush() the output stream,
and then immediately close the BluetoothSocket. About 50% of the time
the other device sees the SessionData message, the other 50% it
doesn't.

If we defer the close() call for 1 second this problem goes away. It
looks like the root problem is that flush() on the OutputStream
obtained via BluetoothSocket.getOutputStream() doesn't work as
intended. Will follow up internally with the Bluetooth team about
this, for now this work-around is needed.

Bug: None
Test: Manually tested, session termination now works reliably with L2CAP