GattServer: Remove assumption about devices connecting and disconnecting.
Our GATT server currently assumes that `onConnectionStateChange` will
only be called for the BLE device we're expecting to communicate with
(e.g. the remote mDL app if using mdoc BLE central client mode) but
this is not true since e.g. already connected peripherals may connect
and disconnect all the time, for example a watch already paired with
the device.

Remove this assumptions and instead set `mCurrentConnection` only when
a device writes 0x01 to the State characterstic to indicate that it is
ready to exchanging data, as per 18013-5 section 8.3.3.1.1.

Also make the debug/log messages more useful by including the MAC
address of the remote device.

A fix for this was proposed in PR #90 but had some issues.

Bug: None
Test: Manually tested using both mdoc BLE central client mode and mdoc BLE peripheral server mode.