BLE: Change requested MTU to 515 and add better debug for chunking.
Android supports MTUs up to 517 which we usually will get if the
communication is between two Android devices. However this means that
the chunks being written to Client2Server and Server2Client
characteristics are 514 bytes long which is in violation of the
Bluetooth spec which says they cannot be bigger than 512 bytes
(Bluetooth Core specification section Part F 3.2.9 Long attribute
values).

Starting with Android 13, it seems like this 512 limit is enforced
which means that our mDL app stopped working when either the mDL or
the mDL reader is on Android 13.

The problem is fixed by simply requesting an MTU of size 515 instead
of 517.

Also add better debug/logging to help catch this and similar problems
and add code to report an error if a non-final chunk isn't the MTU
size less 3, as required by 18013-5.

Bug: None
Tests: mDL and mDL reader apps work on Android 13.