Add better debug for BLE chunking.
Ran into a possible bug with an Android device when using BLE and mdoc
central client mode. The mDL is sending around 16 KiB of response with
a negotiated MTU of 517 meaning that data needs to be chunked into
sizes of MTU - 3 = 514 as per 18013-5.

This change adds debug to make it evident that somehow two bytes of
each write to the Client2Server characteristic are lost
_somewhere_. On the mDL side we get the following output:

 I/GattClient: TRANSPORT: drainWritingQueue false
 I/GattClient: TRANSPORT_VERBOSE: Sending chunk with 514 bytes (last=false)
 I/GattClient: TRANSPORT: onCharacteristicWrite 0 00000006-a123-48ce-896b-4c76973373e6
 [...]
 I/GattClient: TRANSPORT_VERBOSE: Sending chunk with 164 bytes (last=true)
 I/GattClient: TRANSPORT: onCharacteristicWrite 0 00000006-a123-48ce-896b-4c76973373e6
 I/GattClient: TRANSPORT: drainWritingQueue false

however on the mDL reader side we get the following

 I/GattServer: TRANSPORT_VERBOSE: Received chunk with 512 bytes (last=false), incomingMessage.length=511
 I/VerificationHelper: INFO: onError for com.android.identity.DataTransportBleCentralClientMode@e9e9d: java.lang.Error: Invalid size 512 of data written Client2Server characteristic, expected size 514

indicating that only 512 bytes are received. On other Android devices,
514 bytes are received as is expected. This only happens on this
particular Android device with a specific version of the OS. Next step
is to investigate if this is an OS or lower-level bug.

Also add code to report an error if a non-final chunk isn't the MTU
size less 3, as required by 18013-5.

Bug: None
Tests: Manually tested that presentations work.