BLE: Add a setting to clear BLE Service Cache just before service discovery.
We have reports that when using BLE mdoc central client mode with a
particular reader things will work better if we force Android to clear
its internal BLE Service Cache before we initiate service
discovery. This is possible using only unsupported API and
introspection. The function in question is the refresh() method on the
BluetoothGatt class.

To better debug and investigate this, add API to enable this behavior
and surface this setting in both the holder and reader reference
apps. The default value for this setting is off.

Test: Manually tested
Bug: None