Add EngagementGenerator and EngagementParser.
This is mostly just refactoring existing code into Engagement{Parser,
Genenerator} except for newly added support for OriginInfo which is
used in upcoming ISO 18013-7 and ISO 23220-4 standards. This work will
be used in mdoc:// support for over-the-Internet mdoc presentations
according to 18013-7.

Existing QrEngagementHelper and NfcEngagementHelper haven't been wired
up to use this yet as that will include migrating DataRetrievalAddress
classes to the new ConnectionMethod classes. This will happen in an
upcoming change.

Test: New unit tests and all unit tests pass
Bug: None