Rework logging and move some unit tests to the non-Android-specific test suite.
Simplify API surface by removing the concept of logging levels and
instead, just provide a way for applications to globally
enable/disable logging. Introduce Logger class.

To ensure that SessionEncryptionReader, DeviceRequestGenerator, and
DeviceResponseParser can be used on the server side, move unit tests
for these classes (and others) to the test suite that isn't
Android-specific. As an added benefit this test suite runs as part of
presubmit.

Test: All unit tests pass
Test: Manually tested appholder and appreader with a transaction.