Add support for reverse engagement and HTTP transport. (#128)
This adds support in the identity library for reverse engagement and
also support in both reference apps to use the new functionality. To
do this nicely add use a Builder for both VerificationHelper and
PresentationHelper.

Currently the HTTP transport doesn't support TLS. Support for this
will be added in a future commit.

Remove DataRetrievalAddress and DataRetrievalListenerConfiguration
abstractions. Instead just use ConnectionMethod and add new
DataTransportOptions as a way for the application to specify options /
preferences which isn't encoded in a ConnectionMethod (for example
whether to use L2CAP).

Move all NDEF and DeviceRetrievalMethod logic from DataTransport
classes to ConnectionMethod classes.

Test: All unit tests pass
Test: Manually tested both w/ both normal and reverse engagement.