Use HTTP transport for reverse engagement in mdoc reader.
This required a slight refactor of DataTransport such that
listening/connecting transports use the same API and the role is just
passed in at construction time (ROLE_MDOC and ROLE_MDOC_READER).

Specifically for DataTransportHttp when used on the mdoc reader side
it'll return URIs of the form http://<ip>:<port>/mdocreader where <ip>
is the local IP address and <port> is the dynamically chosen TCP port
number returned by the kernel when the listening socket was created.

Test: All unit tests pass.
Test: Manually tested both forward and reverse engagement.
Bug: None