Use correct AAMVA namespace and add new reader option for US transportation.
Also add a couple of mandatory AAMVA data elements in the default
self-signed mDL. The new reader option is called "mDL for US
transportion" and will request an asssorted set of data elements in
both the org.iso.18013.5.1 and org.iso.18013.5.1.aamva namespace.

This actually uncovered a bug in our holder app insofar that two
documents are included in the response, one for each namespace. A
subsequent change will fix this to be just one document.

Test: Manually tested.