Implement feature version 202301.
Feature version 202301 for the platform API is described by the CLs in
the following Gerrit topic:

 https://android-review.googlesource.com/q/topic:ic-android14

The first part of this feature version is support for specifying that
an AuthKey should be replaced if it's going to expire within a certain
amount of time configurable by the application. This also adds a way
for the application to learn about the expiration time of currently
configured AuthKeys. Combined, this allows an application to get a
perfect picture of which AuthKeys are available, when they expire, and
allows the application to refresh AuthKeys well ahead of expiration
dates.

The second part of this feature version is relaxing requirements on
SessionTranscript so the APIs can be used even when mdoc session
encryption isn't being used. This allows the API to be used with other
APIs such as https://github.com/WICG/mobile-document-request-api

Add tests for all this.

A future change will include wiring up the platform API to
HardwareIdentityCredentialStore + friends. This will be done when the
first Android 14 SDK is available.

Test: New unit tests and all unit tests pass