Support referrer not being set for 18013-7 transactions.
Some mobile web browsers (Firefox on Android) do not convey the
referrer URL which makes our application bail. Instead, just attempt
the request and convey an empty OriginInfos array and let the mdoc
reader website deal with an empty OriginInfos.

To implement this, we need to change the API on EngagementGenerator so
it's possible to set an empty OriginInfos because 18013-7 says that
this must be present (even if empty).

Test: Manually tested.