Fix COSE_Key encoding.
RFC 8152 section 13.1.1.  Double Coordinate Curves says to use [SEC1]
for converting X coordinates to bstr (and to preserve leading zero
octects) which implies the octet string for each coordinate is always
32 bytes (for curve P-256).

This was discovered when testing against an implementation that wasn't
as tolerants as other implementations.

Also add a warning if receiving/decoding a COSE_Key where this
requirement isn't met.

Test: New unit test
Test: All unit tests pass
Test: Manually performed a presentation