Rework how NFC APDUs are routed through the application to the library.
When engaging with a NFC reader/writer device which sends multiple
APPLICATION_SELECT for the NDEF application, our mdoc application gets
confused and brings up multiple NfcEngagementHelper and the end result
is that the APDUs are not sent.

Resolve this by removing NfcApduRouter and its listening mechanism
(which is the root cause of this problem) and instead manually push
APDUs from the HostApduService into its NfcEngagementHelper instance.

Test: All unit tests pass
Test: Manually tested with NFC engagement and both BLE modes
Test: Manually tested with NFC engagement and NFC data transfer
Test: Manually tested with QR engagement and NFC data transfer
Test: Manually tested with iOS tag reader (verified correct Handover Select message)