Logging improvements and CBOR helper for printing diagnostics.
Add new CborUtil.toDiagnostics() which returns a string with the given
CBOR (either a DataItem or a byte[]) print in diagnostics format as
per RFC 8949 Section 8. Add flags so the caller can request it to be
pretty-printed (indent and newlines) and also whether it should print
embedded CBOR bytestrings.

Also introduce new Logger methods which prints CBor (both hex and
diagnostics) or Hex for a given byte[], and adjust callsites to use
this. This cleans up callsites since they no longer need to use
Logger.isDebugEnabled() to avoid a costly conversion in the event
that debug isn't enabled.

Test: New unit tests and all unit tests pass