prevent a ble crash due to several scanResult callbacks
DataTransportBleCentralClientMode `onScanResult` is called several times before the `stopScan` takes effect.
Therefore, the `gattClient` is trying to connect to the device again.
The internal implementation of `connectGatt`create a new `BluetoothGatt` instance and later `onConnectionChanged` is called for each instance and later causes a crash.
read more: https://stackoverflow.com/a/38276808/4940838