Send message with status 10 if unable to decrypt message from reader.
Previous behavior is that we would just terminate the session. With
this change before session termination we send a message with no data
and status 10 which according to 18013-5 Table 20 means "Error:
session decryption".

It's not clear the standard actually requires an mdoc to do this but
unfortunately it's something that certain test suites will check for.

Add new unit test to check the behavior and introduce symbolic
constants instead of hardcoding 10 and 20.

Test: New unit test and all unit tests pass.
Test: Manually tested reference apps on two Pixel devices.
Issue: 115