Emit "happy beep" when NFC engagement is complete
This is helpful for holders to know when it's safe to move their device out of the NFC field.

To implement this, we first removed the sound associated with discovering a tag by adding a flag to the NfcAdapter's enableReaderMode function call. Then, to add the desired beep, we used a MediaPlayer which provides flexibility in which noise is played as well as at what point the noise is played.

Test: Manually tested.