Add CredentialInvalidatedException.
This is a runtime exception and it's thrown if the user has removed
the LSKF since the credential was created.

Also update appholder to catch this exception and delete the
credential if caught. Also notify the user of this action with a
toast.

Manually test this behavior by creating a credential w/ user auth when
the device has a LSKF. Then remove the LSKF and try presenting it.

Test: Manually tested
Issue: 247