Refactor to decouple request/response from the rest of IdentityCredential.
This is to make it easier for Server side implementations to dependent on request generation and response parsing.
To do this the android portion of code is now in 'identity-android' while the rest is in 'android'. To make docs work
dokka is being used to generat across both (uses the dokkaHtmlMultiModule command). This can be setup to create javadocs
as well though currently it only is doing html.

This also adds support for overridding the logger using a LoggerPrinter, and implements the android one.