Slightly refactor API to make it possible to properly auto-close connections.
Fix a problem with BLE data transport methods where data isn't flushed
before the connection is closed.

Also rework DeviceRetrievalHelper.SendMessage() so it's possible to
specify the status code in addition or instead of the DeviceResponse
CBOR to send.

Combined, these two fixes makes it possible for an application to send
status 20 (session termination) in addition to DeviceResponse. Use
this to implement the "Auto close connection" preference in the holder
application. Fix up VerificationHelper so looks at both status and
data and it behaves in the expected way when the holder does this.

This has been tested with both mdoc BLE central client mode and mdoc
BLE peripheral server mode and with the case where holder uses "auto
close connection" and not. For the latter, was also tested with either
party terminating the connection both with using status 20 and when
using the transport-specific way to terminate (BLE characteristic).

Add sendTransportSpecificTermination() which can be used instead of
sending status 20. Also remove setSendSessionTerminationMessage()
since the app is now expected to always perform session termination
itself before calling disconnect().

Fixes #270.