Test support for all curves in mdoc session encryption
Combined SessionEncryptionHolder and SessionEncryptionReader into a single SessionEncryption object, differentiated by role. Updated all references to the old classes in the library. Added a testCurve function to ensure reader and holder can encrypt and decrypt messages passed between them with all types of curves, which required additional constants to represent each curve and a change to the createEphemeralKeyPair Util function in order to create Ephemeral keys based on the inputted curve identifier. Edwards curves are not tested yet, but will be added after Keystore abstraction changes.