Add Android storage and keystore implementations.
This adds AndroidStorageEngine which encrypts stored values using a
hardware-backed 128-bit AES key and Galois Counter Mode. Encryption of
data is needed because some older Android devices are not guaranteed
to have disk encryption.

Also adds an Android Keystore-based implementation of KeystoreEngine
as well as a test app for interactively testing things that aren't
possible to test in unit tests, including auth-bound keys.

Test: New unit tests and all unit tests pass
Test: Manually tested using new test app