Add new credential store.
This adds a new credential store, replacing IdentityCredentialStore.

The new credential store is built on top of StorageEngine and
KeystoreEngine introduced in previous commits and as such can work in
non-Android environments. Additionally, the credential store is
generic in the sense that it is not tied to ISO/IEC 18013-5:2021 at
all, it can work for any credential shape, regardless of data-model,
presentation-protocol, or issuance-protocol.

A future commit will add ways to use this new credential store for
18013-5 presentations and also migrate credentials from
IdentityCredentialStore, preserving key material and PII previously
issued.

Test: New unit test and all unit tests pass.