Make PresentationSession optional.
With PresentationSession now in the 'legacy' package it would be
helpful if an application could just pass a KeyPair and PublicKey
which can be used for mdoc session authentication.

Also remove EC Curve constants from Constants class since these are
already available in KeystoreEngine interface

Also add docs for NfcEngagementHelper and QrEngagementHelper.

Test: All unit tests pass.
Test: Manully tested.