Rename KeystoreEngine to SecureArea.
It's likely we want to add features implemented in the secure area in the
future that are specific to real-world identity credentials in addition to
symmetric and asymmetric keys that are guaranteed to never leave the secure
area. For example, this could include keys and objects needed for ZKP and
storage of credential data.

Also add a feature to the Android Keystore implementation of SecureArea
to allow specifying whether LSKF, biometric, or both can be used to unlock
the key. Make a note in the docs that this only works with API 30 or later.

Also fix a bug in DeviceRetrievalHelperTest.

Test: All unit tests pass.
Test: Manually tested with holder and reader.