Rework BouncyCastleSecureArea.
First of all, rename this to SoftwareSecureArea since there is no need
to leak the fact that we're currently using BouncyCastle which could
change in the future.

Instead of always using self-signed keys, make it possible to specify
the attestation key to use at key creation time. Without this we
cannot easily support curves which cannot be used for signing (for
example X25519). Modify holder app to create an attestation root
on demand.

Introduce attestation extension so SoftwareSecureArea keys also
can convey the attestation challenge. The OID for this has been
reserved (it's 1.3.6.1.4.1.11129.2.1.39) and we can extend this
if needed. Add code for dealing with this and use it in
SoftwareSecureArea.

Show the curve of DeviceKey in the reader app.

Make sure we use the BouncyCastle library bundled with the app
instead of the Conscrypt-based implementation that may come
with the OS. Do this for both wallet and reader app.

This has been manually test with mdocs using both ECDSA and MAC
authentication with all curves except Ed25519, X25519, Ed448,
and X448. These curves still have some serialization problems,
we'll revisit this in a future PR.

The app still uses the name "Bouncy Castle", we'll address that
in a future PR.

Also update to the latest available BouncyCastle, version 1.70.

Test: Manually tested.
Test: All unit tests pass.