Make Credman presentations faster and smoother.
Turns out that logging CBOR is really really slow (filed as Issue #387)
so disable that. This removes a 2-3 second pause in the reader when
requesting a credential via Credman.

Also add

 android:theme="@android:style/Theme.Translucent.NoTitleBar.Fullscreen"

to the GetCredentialActivity in the manifest so the BiometricPrompt
appears on top of the requesting app. This makes the whole thing look
100x better since there's no annoying flash.

Test: Manually tested
Signed-off-by: David Zeuthen <zeuthen@google.com>