Add support for per-MSO data elements.
Our current model involves sending all PII to the device and then
a bunch of MSOs + associated data (StaticAuthData) where the
latter doesn't include any PII at all. Add the ability for
StaticAuthData to also include PII meaning the issuer has the
option to send none, some, or all PII along with each MSO.

Also use this in appholder by embedding the string "MSO <counter>" on
top of the portrait image where <counter> is the count of the MSO
and increasing for every use. In this example, only the portrait
data element is per-MSO.

Fixes #405.

Test: Manually tested.
Test: New unit tests and all unit tests pass.