Allow instantiating SecureArea.CreateKeySettings object.
In order to support new kinds of SecureArea without app changes,
change the SecureArea.CreateKeySettings class so it can be
instantiated by apps.

Also stop carrying the SecureArea class name in CreateKeySettings
subclasses, instead have users of CreateKeySettings take a separate
SecureArea class. Make it explicit that different SecureArea instances
can be used by CredentialKey and even for different Authentication
Keys.

Introduce the concept of an "identifier" for a SecureArea instead
of Credential, Credential.Authentiationkey, and
Credential.PendingAuthenticationKey classes all using the Java class
name when persisting on-disk. This paves the way to have multiple
SecureArea instances of the same type but with different configuration.
We'll need this for CloudSecureArea so an app can have two instances
pointing to different providers, for different credentials. Also
add "display name" since that is needed for Issue #380.

Unfortunately this breaks the on-disk format but this is going
to happen _anyway_ as we're renaming the project and moving all
the code to new packages as part of the contribution of the
code to the OpenWallet Foundation. In the near future we'll start
committing to stable on-disk formats. For now, just update
the directory used for storing data in appholder to a new
location to avoid breaking existing installations.

Test: All unit tests pass.
Test: Manually tested